Chaskiel Grundman: This patch fixes some broken PTS behavior that resulted from two misconceptions that I had about PTS:
1) that the PTS response from the card must be an exact reflection of the 
request in order to be considered successful. 7816-3 says that the card 
may refuse the PTS1 byte by sending a response that doesn't include PTS1.
2) that the PTS2 byte was an N (guard time) value. PTS2 is in fact a 
coded-value, of which only the low 2 bits are defined.

The patch adds a new function, ifd_verify_pts, that validates a pts 
response according to the pts request that would have been sent for a 
given atr_info. it clears the TA1 field from the atr_info if PTS1 was 
refused by the card.