Fix OpenBSD issues.
fixes sys-bsd.c to use the correct device
filenames on OpenBSD. An earlier patch already fixed this for the control
endpoint, but this caused the other endpoints to be accessed under
'/dev/ugenX.00.E'. This patch causes the correct filenames to be used
for all endpoints.

Also, Fix BSD support of USB devices.

By David Imhoff and Hasso Tepper.